// apps/client/src/audio-samples.ts
// Generated by scripts/prepare-audio.py from licensed recordings.
export const AUDIO_SAMPLES = {
  diceRattle: {
    url: '/audio/sfx/diceRattle.ce38c2b7f8d1.wav',
    duration: 0.22,
  },
  diceContact: {
    url: '/audio/sfx/diceContact.322047d5cee0.wav',
    duration: 0.19002267573696144,
  },
  wood: {
    url: '/audio/sfx/wood.42c59f1648b4.wav',
    duration: 0.23002267573696145,
  },
  woodHeavy: {
    url: '/audio/sfx/woodHeavy.ee6d3defeea0.wav',
    duration: 0.28,
  },
  plank: {
    url: '/audio/sfx/plank.3484750ecc99.wav',
    duration: 0.3,
  },
  stone: {
    url: '/audio/sfx/stone.cfb7638b1d47.wav',
    duration: 0.5100226757369615,
  },
  paperSlide: {
    url: '/audio/sfx/paperSlide.40fbd4665fba.wav',
    duration: 0.22,
  },
  paperPlace: {
    url: '/audio/sfx/paperPlace.ae3d7fa14c7a.wav',
    duration: 0.24997732426303854,
  },
  paperFan: {
    url: '/audio/sfx/paperFan.86e76d607f4a.wav',
    duration: 0.6300226757369615,
  },
  cloth: {
    url: '/audio/sfx/cloth.341d9358b03a.wav',
    duration: 0.5100226757369615,
  },
  steel: {
    url: '/audio/sfx/steel.be30d82d1190.wav',
    duration: 0.30970521541950113,
  },
  turn: {
    url: '/audio/sfx/turn.942415719b3e.m4a',
    duration: 1.3239909297052155,
  },
  award: {
    url: '/audio/sfx/award.028971f8777e.m4a',
    duration: 1.0019954648526077,
  },
  magic: {
    url: '/audio/sfx/magic.15d0551e7124.m4a',
    duration: 0.9557369614512472,
  },
  join: {
    url: '/audio/sfx/join.f171e2253d5a.m4a',
    duration: 0.4939909297052154,
  },
} as const;